+++ src/NORS.Plugin/NorsPlugin.cs
@@ -5,7 +5,7 @@ using UnityEngine;
 
 namespace NORS.Plugin
 {
-    [BepInPlugin(Guid, "Nuclear Option Radio System", "0.7.3")]
+    [BepInPlugin(Guid, "Nuclear Option Radio System", "0.7.4")]
     public class NorsPlugin : BaseUnityPlugin
     {
         public const string Guid = "com.dsr.nors";
@@ -26,7 +26,7 @@ namespace NORS.Plugin
             DontDestroyOnLoad(_hub);
             _hub.AddComponent<NorsHub>();
 
-            Log.LogInfo("Nuclear Option Radio System v0.7.3 online. Press " +
+            Log.LogInfo("Nuclear Option Radio System v0.7.4 online. Press " +
                         NorsConfig.PanelKey.Value + " for the radio panel.");
         }
     }